pyproject.toml
[tool.poetry]
name = "mini-argos-poc"
version = "0.1.0"
description = "A proof-of-concept for a mini-ARGOS system with agent collaboration."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
google-adk = ">=0.1.0"
agent-starter-pack = ">=0.1.0"
redis = ">=5.0.0"
fastapi = ">=0.104.0"
uvicorn = {extras = ["standard"], version = ">=0.24.0"}
copilotkit = ">=0.1.0"
google-cloud-speech = ">=2.21.0"
google-cloud-texttospeech = ">=2.14.0"
websockets = ">=12.0"
pypdf2 = ">=3.0.0"
pdfplumber = ">=0.10.0"
arxiv = ">=2.0.0"
networkx = ">=3.2"
matplotlib = ">=3.8.0"
scikit-learn = ">=1.3.0"
dspy-ai = ">=2.0.0" # For DSPy-powered task decomposition
tavily = ">=0.0.8"
python-dotenv = ">=1.0.0"
mcp = {extras = ["cli"], version = ">=0.1.0"}

[tool.poetry.dev-dependencies]
pytest = "^7.4"
python-dotenv = ">=1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
